fix https://github.com/cucumber/cucumber-ruby/issues/901
to make it 2 orders of magnitude faster two things were needed:
1. do not try all locations, only those with matching file name
2. do not order test cases after the fact, put them in proper place from the beginning

The above also removes duplicates for good or for bad.

This is achieved by adding a Test::Case method to return index of matching location so we don't need to figure out later which one exactly the matching location was.